test(providers): consolidate registry tests, split out interface tests
provider_test.go and registry_test.go both covered the registry:
provider_test.go through the ProviderRegistry methods, registry_test.go
through the package-level wrappers. Move the seven method and
DefaultRegistry tests into registry_test.go so all registry coverage
lives in one file, and rename the remainder to interfaces_test.go, which
now holds the two interface tests and the mock types they need.

No test was merged away. The wrappers are not pass-throughs, so a test
through RegisterProvider covers lines a test through the method does not:
the method returns an error on a duplicate name, while the wrapper
swallows that error and logs it. Both paths keep their own test.

Pure move. No test logic, assertion, or name changed. Closes #50.
